make a separate non-ui target
this will allow another UI (like CarPlay) to depend upon StadiaMapsAutocompleteSearchAPI w/o SwiftUI
